fastapi/requirements.txt
fastapi==0.115.5
uvicorn[standard]>=0.34.0
python-multipart==0.0.18
yt-dlp==2024.11.18
requests==2.32.3
pydantic==2.10.3
python-dotenv==1.0.1
google-adk==1.0.0
google-genai>=1.14.0
google-cloud-firestore>=2.15.0
google-cloud-storage>=2.16.0
google-cloud-speech>=3.30.0
google-cloud-texttospeech>=2.16.0
firebase-admin>=6.5.0
tenacity==8.2.3
redis==5.0.4
rq==1.16.2
pusher==3.3.2
ffmpeg-python==0.2.0
motor==3.6.0
pymongo==4.10.1
httpx==0.27.2
